+++ validationframework-core/src/main/java/com/google/code/validationframework/base/rule/OrCompositeBooleanRule.java
@@ -25,6 +25,7 @@
 
 package com.google.code.validationframework.base.rule;
 
+import com.google.code.validationframework.api.common.Disposable;
 import com.google.code.validationframework.api.rule.Rule;
 import com.google.code.validationframework.base.transform.OrBooleanAggregator;
 import com.google.code.validationframework.base.transform.Transformer;
@@ -80,6 +81,23 @@ public class OrCompositeBooleanRule<RI> extends AbstractCompositeRule<RI, Boolea
         }
 
         // Aggregate results
-        return aggregator.transform(results);
+        Boolean aggregated = null;
+        if (aggregator != null) {
+            aggregated = aggregator.transform(results);
+        }
+
+        return aggregated;
+    }
+
+    /**
+     * @see AbstractCompositeRule#dispose()
+     */
+    @Override
+    public void dispose() {
+        super.dispose();
+
+        if (aggregator instanceof Disposable) {
+            ((Disposable) aggregator).dispose();
+        }
     }
 }